backend/requirements.txt
Flask
flask-cors
flask_sqlalchemy
bcrypt
PyJWT
openai
requests>=2.25.0
pathlib
python-dotenv
rasa>=1.10.0